examples/my-ai-service/pyproject.toml
[project]
name = "my_ai_service"
version = "0.1.0"
description = "Generated FastAPI service via IDP-Lite"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.110.0",
  "uvicorn>=0.24.0",
  "pydantic>=2.6.0",
  "pyyaml>=6.0",
  "structlog>=23.1.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=7.4.0",
  "httpx>=0.27.0",
  "ruff>=0.6.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
filterwarnings = ["ignore::DeprecationWarning"]

[tool.ruff]
line-length = 100
select = ["E", "F", "I", "W"]
